backend/requirements.txt
argcomplete==3.1.2
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.1.0
billiard==4.1.0
build==0.10.0
CacheControl==0.12.11
cachetools==5.3.1
certifi==2023.5.7
cfgv==3.4.0
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
commitizen==3.10.0
crashtest==0.4.1
decli==0.6.1
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.6
Django==4.2.5
django-filter==23.2
django-import-export==3.3.1
djangorestframework==3.14.0
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.9.1
drf-standardized-errors==0.12.5
dulwich==0.21.5
et-xmlfile==1.1.0
filelock==3.12.0
flake8==6.0.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
gunicorn==20.1.0
html2text==2020.1.16
html5lib==1.1
httplib2==0.22.0
identify==2.5.29
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
iniconfig==2.0.0
installer==0.7.0
isort==5.12.0
jaraco.classes==3.2.3
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
MarkupSafe==2.1.3
mccabe==0.7.0
more-itertools==9.1.0
msgpack==1.0.5
nodeenv==1.8.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==2.6.2
pluggy==0.13.1
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.1
pre-commit==3.4.0
prompt-toolkit==3.0.36
protobuf==4.24.0
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.1.1
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==6.2.4
pytest-django==4.4.0
python-dateutil==2.8.2
pytz==2023.3
pywin32-ctypes==0.2.0
PyYAML==6.0.1
questionary==2.0.1
rapidfuzz==2.15.1
requests==2.30.0
requests-toolbelt==0.10.1
rsa==4.9
setuptools-git==1.2
shellingham==1.5.0.post1
six==1.16.0
sqlparse==0.4.4
structlog==23.1.0
tablib==3.5.0
termcolor==2.3.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
trove-classifiers==2023.5.2
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
virtualenv==20.21.1
wcwidth==0.2.6
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.15.0
